requirements.txt
python-telegram-bot==20.3
PyPDF2==3.0.1
google-generativeai==0.1.0
Pillow==9.5.0